STGPKP

SpatiotemporalGPs.STGPKF.STGPKFProblem — Method

STGPKFProblem(pts, ks, kt, ΔT)

Defines a spatiotemporal Gaussian Process Kalman Filter problem. Parameters are:

pts: grid points, a vector of all points. Ideally, eltype(pts) should be StaticVector for efficiency

ks: spatial kernel, must be of type AbstractKernel

kt: temporal kernel, must be of type AbstractKernel (but only AbstractMaternKernel is implemented)

ΔT: sampling period

SpatiotemporalGPs.STGPKF.Matern — Method

Matern(order, σ, l)

creates a Matern kernel with order order, variance σ, and lengthscale l. Order must be (1/2, 3/2, or 5/2). Returns a Matern12, Matern32, or Matern52 kernel.

SpatiotemporalGPs.STGPKF.SquaredExponential — Method

SquaredExponential(σ, l)

creates a Squared Exponential kernel with variance σ and lengthscale l. Returns a SqExp kernel.

SpatiotemporalGPs.STGPKF.checkdims — Method

checkdims(prob, state)

checks that the dimensions of the state and the problem match

SpatiotemporalGPs.STGPKF.get_estimate — Method

get_estimate(problem, state)

returns the estimate of the Kalman Filter for all grid points, in a Vector{F} format. The outer vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.get_estimate_clarity — Method

get_estimate_clarities(problem, state)

returns the clarity of the estimated spatiotemporal field at all grid points, in a Vector{F} format. The vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.get_estimate_covariance — Method

get_estimate_covariance(problem, state)

returns the kalman filter's covariance of the estimated spatiotemporal field at all grid points, in a Vector{F} format. The vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.get_estimate_percentile — Method

get_estimate_percentile(problem, state, percentile)

returns the percentile-% quantile of the estimated spatiotemporal field at all grid points, in a Vector{F} format. The vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.get_estimate_std — Method

get_estimate_std(problem, state)

returns the standard deviation of the estimated spatiotemporal field at all grid points, in a Vector{F} format. The vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.get_marginal_states — Method

get_marginal_states(problem, state)

returns the marginal states of the Kalman Filter for all grid points, in a Vector{KFState} format. The outer vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.get_states — Method

get_states(problem, state)

returns the states of the Kalman Filter for all grid points, in a Vector{SVector{F}} format. The outer vector has same length as problem.pts.

SpatiotemporalGPs.STGPKF.kernel_matrix — Method

kernel_matrix(kernel, X, Y)

Compute the kernel matrix between two sets of points X and Y using the kernel function kernel. X must be a vector of points Y must be a vector of points

SpatiotemporalGPs.STGPKF.quantile — Method

quantile(μ, σ, q)

For a normal distribution with mean μ and standard deviation σ, this function returns the q-th quantile.

SpatiotemporalGPs.STGPKF.stgpkf_correct — Method

stgpkf_correct(prob, state, pt, y, σ_m)

corrects the state of the Kalman Filter given a single point measurement at $pt$ with value $y$ and measurement noise standard deviation $σ_m$.

SpatiotemporalGPs.STGPKF.stgpkf_correct — Method

stgpkf_correct(prob, state, pts, ys, Σm)

corrects the state of the Kalman Filter given multiple point measurements at $pts$ with values $ys$ and measurement noise covariance matrix $Σm$.

SpatiotemporalGPs.STGPKF.stgpkf_initialize — Method

stgpkf_initialize(problem)

returns a KFState that represents the initial state of the Kalman Filter for all grid points

SpatiotemporalGPs.STGPKF.stgpkf_predict — Method

stgpkf_predict(prob, state)

predicts the next state of the Kalman Filter for all grid points